+++ src/main/java/org/example/siljeun/domain/seatscheduleinfo/repository/SeatScheduleInfoRepository.java
@@ -1,7 +1,7 @@
 package org.example.siljeun.domain.seatscheduleinfo.repository;
 
 import org.example.siljeun.domain.schedule.entity.Schedule;
-import org.example.siljeun.domain.seat.entity.SeatScheduleInfo;
+import org.example.siljeun.domain.seatscheduleinfo.entity.SeatScheduleInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
