+++ src/main/java/org/example/siljeun/domain/seatscheduleinfo/scheduler/SeatStatusPreloaderScheduler.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.example.siljeun.domain.schedule.entity.Schedule;
 import org.example.siljeun.domain.schedule.repository.ScheduleRepository;
 import org.example.siljeun.domain.seatscheduleinfo.repository.SeatScheduleInfoRepository;
-import org.example.siljeun.domain.seat.entity.SeatScheduleInfo;
+import org.example.siljeun.domain.seatscheduleinfo.entity.SeatScheduleInfo;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
