+++ src/main/java/org/example/siljeun/domain/seatscheduleinfo/service/SeatScheduleInfoService.java
@@ -6,7 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.example.siljeun.domain.schedule.entity.Schedule;
 import org.example.siljeun.domain.schedule.repository.ScheduleRepository;
 import org.example.siljeun.domain.seatscheduleinfo.repository.SeatScheduleInfoRepository;
-import org.example.siljeun.domain.seat.entity.SeatScheduleInfo;
+import org.example.siljeun.domain.seatscheduleinfo.entity.SeatScheduleInfo;
 import org.example.siljeun.domain.seat.enums.SeatStatus;
 import org.example.siljeun.global.lock.DistributedLock;
 import org.springframework.data.redis.core.RedisTemplate;
